Use bfloat16 for CUDA devices for numerical stability
Float16 can cause NaN probability issues with some models.
Bfloat16 has better numerical stability for LLM inference.

+++ wisent/core/models/wisent_model.py
@@ -93,7 +93,7 @@ class WisentModel:
             load_kwargs["torch_dtype"] = torch.float16
             load_kwargs["device_map"] = "mps"
         elif self.device == "cuda" or self.device == "auto":
-            load_kwargs["torch_dtype"] = torch.float16
+            load_kwargs["torch_dtype"] = torch.bfloat16  # More stable than float16
             load_kwargs["device_map"] = "auto"
         else:
             load_kwargs["torch_dtype"] = torch.float32
